fix: wire the modules Boot 4 stopped auto-configuring
Compiling was never the hard part. Boot 4 stopped auto-configuring a technology
just because its library is on the classpath, and every consequence of that is
silent:

- Flyway. flyway-core without spring-boot-starter-flyway starts the application,
  applies no migrations at all, and fails on the first missing table. 239 test
  errors reading "missing table [alert_events]" is what that looks like.
- WebClient. Its auto-configuration left the WebFlux starter, which is now only
  the reactive server stack. The API only ever wanted WebClient from it, so it
  takes spring-boot-starter-webclient instead; the worker keeps both, because
  webflux is what serves the actuator port its healthchecks read.
- Kafka, whose auto-configuration now arrives only with its own starter.
- Redisson 3.52 wires against Boot 3's module layout. 4.7.0 also renamed
  RScript.ReturnType's constants to the Java types they return - INTEGER to
  LONG, MULTI to LIST - which the circuit breaker's and ordering buffer's Lua
  scripts name directly.

Jackson is the one with a judgement in it. Boot 4 serializes HTTP through
Jackson 3, and two DTOs put a Jackson 2 JsonNode on the wire; one of them,
PlanResponse.features, is a JSONB column on the Plan entity that the
schema-validation gate also reads. Moving to Jackson 3 is therefore an entity
change with a migration behind it - not something an upgrade whose purpose is
closing five CVSS 9.8s should carry along. So HTTP stays on Jackson 2 via
spring-boot-jackson2 and preferred-json-mapper, both deprecated or easily
mistaken for decoration.

The two halves of that fail differently, which is why the new test covers both
directions: without the property, *reading* a body into a JsonNode field
answers 500 while writing one still works. EventIngestRequest.data is such a
field, so what breaks first is every event the platform ingests. A first draft
of this test checked only responses and "proved" the property was dead
configuration - it was not.

springdoc moves to the 3.x line for Boot 4. OpenApiDriftIntegrationTest passes
without regenerating anything, so the published contract is unchanged.

Verified where the tests cannot reach: the integration tests exclude Redisson's
auto-configuration and mock every service that touches Redis, so this was run
against a live stack. Both services healthy on 4.1.1, Redisson 4.7.0 connected,
migrations applied, and an event ingested through to a delivery that succeeded
on its first attempt - with a deliberately unreachable endpoint alongside it
climbing the retry ladder as designed.

+++ webhook-platform-worker/src/test/java/com/webhook/platform/worker/service/OrderingBufferServiceTest.java
@@ -91,7 +91,7 @@ class OrderingBufferServiceTest {
         // Faithfully mirrors lua/ordering_cursor_cas.lua: SET only if newVal > current
         // (missing key treated as "current == false"), atomically (synchronized -- this test
         // fake is what stands in for Redis's single-threaded Lua execution guarantee).
-        when(rScript.eval(eq(RScript.Mode.READ_WRITE), any(String.class), eq(RScript.ReturnType.INTEGER),
+        when(rScript.eval(eq(RScript.Mode.READ_WRITE), any(String.class), eq(RScript.ReturnType.LONG),
                 any(List.class), any(), any())).thenAnswer(invocation -> {
             List<String> keys = invocation.getArgument(3);
             String key = keys.get(0);
